+++ src/wifi/model/eht/eht-configuration.cc
@@ -36,6 +36,21 @@ NS_LOG_COMPONENT_DEFINE("EhtConfiguration");
 
 NS_OBJECT_ENSURE_REGISTERED(EhtConfiguration);
 
+std::ostream&
+operator<<(std::ostream& os, WifiTidToLinkMappingNegSupport negsupport)
+{
+    switch (negsupport)
+    {
+    case WifiTidToLinkMappingNegSupport::NOT_SUPPORTED:
+        return os << "NOT_SUPPORTED";
+    case WifiTidToLinkMappingNegSupport::SAME_LINK_SET:
+        return os << "SAME_LINK_SET";
+    case WifiTidToLinkMappingNegSupport::ANY_LINK_SET:
+        return os << "ANY_LINK_SET";
+    };
+    return os << "UNKNOWN(" << static_cast<uint32_t>(negsupport) << ")";
+}
+
 EhtConfiguration::EhtConfiguration()
 {
     NS_LOG_FUNCTION(this);
@@ -94,19 +109,17 @@ EhtConfiguration::GetTypeId()
                 UintegerValue(DEFAULT_MSD_MAX_N_TXOPS),
                 MakeUintegerAccessor(&EhtConfiguration::m_msdMaxNTxops),
                 MakeUintegerChecker<uint8_t>(0, 15))
-            .AddAttribute(
-                "TidToLinkMappingNegSupport",
-                "TID-to-Link Mapping Negotiation Support.",
-                EnumValue(WifiTidToLinkMappingNegSupport::WIFI_TID_TO_LINK_MAPPING_ANY_LINK_SET),
-                MakeEnumAccessor<WifiTidToLinkMappingNegSupport>(
-                    &EhtConfiguration::m_tidLinkMappingSupport),
-                MakeEnumChecker(
-                    WifiTidToLinkMappingNegSupport::WIFI_TID_TO_LINK_MAPPING_NOT_SUPPORTED,
-                    "NOT_SUPPORTED",
-                    WifiTidToLinkMappingNegSupport::WIFI_TID_TO_LINK_MAPPING_SAME_LINK_SET,
-                    "SAME_LINK_SET",
-                    WifiTidToLinkMappingNegSupport::WIFI_TID_TO_LINK_MAPPING_ANY_LINK_SET,
-                    "ANY_LINK_SET"))
+            .AddAttribute("TidToLinkMappingNegSupport",
+                          "TID-to-Link Mapping Negotiation Support.",
+                          EnumValue(WifiTidToLinkMappingNegSupport::ANY_LINK_SET),
+                          MakeEnumAccessor<WifiTidToLinkMappingNegSupport>(
+                              &EhtConfiguration::m_tidLinkMappingSupport),
+                          MakeEnumChecker(WifiTidToLinkMappingNegSupport::NOT_SUPPORTED,
+                                          "NOT_SUPPORTED",
+                                          WifiTidToLinkMappingNegSupport::SAME_LINK_SET,
+                                          "SAME_LINK_SET",
+                                          WifiTidToLinkMappingNegSupport::ANY_LINK_SET,
+                                          "ANY_LINK_SET"))
             .AddAttribute(
                 "TidToLinkMappingDl",
                 "A list-of-TIDs-indexed map of the list of links where the TIDs are mapped to "