+++ src/wifi/model/eht/eht-configuration.h
@@ -21,6 +21,7 @@
 #ifndef EHT_CONFIGURATION_H
 #define EHT_CONFIGURATION_H
 
+#include "ns3/deprecated.h"
 #include "ns3/nstime.h"
 #include "ns3/object.h"
 #include "ns3/wifi-utils.h"
@@ -43,12 +44,31 @@ static constexpr uint8_t DEFAULT_MSD_MAX_N_TXOPS = 1;
 /**
  * \brief TID-to-Link Mapping Negotiation Support
  */
-enum WifiTidToLinkMappingNegSupport : uint8_t
+enum class WifiTidToLinkMappingNegSupport : uint8_t
 {
-    WIFI_TID_TO_LINK_MAPPING_NOT_SUPPORTED = 0,
-    WIFI_TID_TO_LINK_MAPPING_SAME_LINK_SET = 1,
-    WIFI_TID_TO_LINK_MAPPING_ANY_LINK_SET = 3
+    NOT_SUPPORTED = 0,
+    SAME_LINK_SET = 1,
+    ANY_LINK_SET = 3
 };
+
+NS_DEPRECATED_3_42("Use WifiTidToLinkMappingNegSupport::NOT_SUPPORTED instead")
+static constexpr auto WIFI_TID_TO_LINK_MAPPING_NOT_SUPPORTED = WifiTidToLinkMappingNegSupport::
+    NOT_SUPPORTED; //!< \deprecated See WifiTidToLinkMappingNegSupport::NOT_SUPPORTED
+NS_DEPRECATED_3_42("Use WifiTidToLinkMappingNegSupport::SAME_LINK_SET instead")
+static constexpr auto WIFI_TID_TO_LINK_MAPPING_SAME_LINK_SET = WifiTidToLinkMappingNegSupport::
+    SAME_LINK_SET; //!< \deprecated See WifiTidToLinkMappingNegSupport::SAME_LINK_SET
+NS_DEPRECATED_3_42("Use WifiTidToLinkMappingNegSupport::ANY_LINK_SET instead")
+static constexpr auto WIFI_TID_TO_LINK_MAPPING_ANY_LINK_SET =
+    WifiTidToLinkMappingNegSupport::ANY_LINK_SET; //!< \deprecated See
+                                                  //!< WifiTidToLinkMappingNegSupport::ANY_LINK_SET
+
+/**
+ * \brief Stream insertion operator.
+ * \param [in] os The reference to the output stream.
+ * \param [in] negsupport The WifiTidToLinkMappingNegSupport.
+ * \return The reference to the output stream.
+ */
+std::ostream& operator<<(std::ostream& os, WifiTidToLinkMappingNegSupport negsupport);
 
 /**
  * \brief EHT configuration